ADD a GroupingRules class
