+++ src/test/java/com/telefonica/iot/cygnus/interceptors/GroupingInterceptorTest.java
@@ -18,11 +18,11 @@
 
 package com.telefonica.iot.cygnus.interceptors;
 
+import com.telefonica.iot.cygnus.interceptors.GroupingRules.GroupingRule;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.flume.event.EventBuilder;
 import org.apache.flume.Event;
-import com.telefonica.iot.cygnus.interceptors.GroupingInterceptor.GroupingRule;
 import com.telefonica.iot.cygnus.utils.Constants;
 import java.io.File;
 import java.io.PrintWriter;
@@ -207,7 +207,7 @@ public class GroupingInterceptorTest {
         System.out.println("Testing GroupingInterceptor.initialize");
         groupingInterceptor = new GroupingInterceptor(groupingRulesFileName);
         groupingInterceptor.initialize();
-        LinkedList<GroupingRule> groupingRules = groupingInterceptor.getGroupingRules();
+        LinkedList<GroupingRule> groupingRules = groupingInterceptor.getGroupingRules().getRules();
         assertTrue(groupingRules.size() == 3); // there are 5 rules, but two are invalid
         GroupingRule firstRule = groupingRules.get(0);
         assertEquals(1, firstRule.getId());
